Правки по замечаниям

+++ src/test/java/CreateUserTest.java
@@ -1,3 +1,4 @@
+import io.qameta.allure.Description;
 import methods.RequestSpec;
 import org.junit.After;
 import org.junit.Before;
@@ -13,6 +14,7 @@ import methods.UserRequests;
 import io.qameta.allure.Step;
 
 import static constants.ApiConstants.BURGERS_URL;
+import static org.apache.http.HttpStatus.SC_FORBIDDEN;
 import static org.hamcrest.Matchers.equalTo;
 
 @RunWith(Parameterized.class)
@@ -45,6 +47,7 @@ public class CreateUserTest {
         RestAssured.requestSpecification = RequestSpec.requestSpecification();
         user = new User(email, password, name);
         userRequests = new UserRequests();
+        accessToken = null; // Инициализация
     }
 
     @Test
@@ -52,14 +55,42 @@ public class CreateUserTest {
     @DisplayName("Проверка создания пользователя")
     public void createUser() {
         Response responseCreate = userRequests.createUser(user);
-        responseCreate.then().statusCode(200).assertThat().body("success", equalTo(true));
-        accessToken = responseCreate.then().log().all().extract().path("accessToken");
-        System.out.println(accessToken);
+        responseCreate.then()
+                .statusCode(200)
+                .assertThat()
+                .body("success", equalTo(true));
+        accessToken = responseCreate.then().extract().path("accessToken");
+    }
+
+    @Test
+    @DisplayName("Попытка создания пользователя с существующим email")
+    @Description("API должно возвращать ошибку 403 при попытке создать пользователя с уже занятым email")
+    public void testCreateUserWithDuplicateEmail() {
+        // Сначала создаем пользователя
+        Response firstResponse = userRequests.createUser(user);
+        accessToken = firstResponse.then().extract().path("accessToken");
+
+        // Пытаемся создать такого же пользователя еще раз
+        Response duplicateResponse = userRequests.createUser(user);
+
+        // Проверяем ответ сервера
+        duplicateResponse.then()
+                .statusCode(SC_FORBIDDEN) // Ожидаем код 403
+                .body("success", equalTo(false))
+                .body("message", equalTo("User already exists"));
     }
 
     @After
     public void deleteUser() {
-        Response responseDelete = userRequests.deleteUser(accessToken);
-        responseDelete.then().log().all().statusCode(202).body("success", equalTo(true));
+        if (accessToken != null) {
+            try {
+                Response responseDelete = userRequests.deleteUser(accessToken);
+                responseDelete.then()
+                        .statusCode(202)
+                        .body("success", equalTo(true));
+            } catch (Exception e) {
+                System.err.println("Ошибка при удалении пользователя: " + e.getMessage());
+            }
+        }
     }
 }
